requirements.txt
Django==5.0.3
django-allauth==0.61.1
django-extensions==3.2.3
django-filter==23.5
django-htmx==1.17.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1